Update Selenium.WebDriver to 4.20.0
Along with this, use the new timeout capability properties. Fixed an issue that the timeouts could be floating point numbers.

+++ src/FlaUI.WebDriver/TimeoutsConfiguration.cs
@@ -5,10 +5,10 @@ namespace FlaUI.WebDriver
     public class TimeoutsConfiguration
     {
         [JsonPropertyName("script")]
-        public int? ScriptTimeoutMs { get; set; } = 30000;
+        public double? ScriptTimeoutMs { get; set; } = 30000;
         [JsonPropertyName("pageLoad")]
-        public int PageLoadTimeoutMs { get; set; } = 300000;
+        public double PageLoadTimeoutMs { get; set; } = 300000;
         [JsonPropertyName("implicit")]
-        public int ImplicitWaitTimeoutMs { get; set; } = 0;
+        public double ImplicitWaitTimeoutMs { get; set; } = 0;
     }
 }
